+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# Changes in version 2.?.? - 2017-??-??
+
+ * Medium changes
+   - Add new optional "build_revision" field to index.json with the
+     Git revision supplied by the calling software.
+
+
 # Changes in version 2.1.0 - 2017-09-15
 
  * Major changes